Various compiler warning fixes

+++ panda/src/pgraph/nodePathCollection.h
@@ -45,8 +45,8 @@ PUBLISHED:
   int get_num_paths() const;
   NodePath get_path(int index) const;
   MAKE_SEQ(get_paths, get_num_paths, get_path);
-  NodePath operator [] (int index) const;
-  int size() const;
+  NodePath operator [] (size_t index) const;
+  size_t size() const;
   INLINE void operator += (const NodePathCollection &other);
   INLINE NodePathCollection operator + (const NodePathCollection &other) const;
